Move graphics handling from shifter to screen.
