+++ shifter.c
@@ -15,18 +15,10 @@
  * 16 MHz	?
  */
 
-#include <stdio.h>
-#include <stdlib.h>
-#include <unistd.h>
-#include <SDL.h>
-#include <fcntl.h>
-#include <sys/time.h>
 #include "common.h"
 #include "cpu.h"
-#include "mfp.h"
 #include "screen.h"
 #include "mmu.h"
-#include "ram.h"
 #include "state.h"
 #include "diag.h"
 #include "glue.h"
@@ -130,14 +122,6 @@ static int vblpre = 0;
 static int vblscr = 0;
 static int hblpre = 0;
 static int hblscr = 0;
-static int framecnt;
-static int64_t last_framecnt_usec;
-static int usecs_per_framecnt_interval = 1;
-
-static int ppm_fd;
-
-static int vbl_triggered = 0;
-static int hbl_triggered = 0;
 
 HANDLE_DIAGNOSTICS(shifter)
 
@@ -277,32 +261,6 @@ static void shifter_state_restore(struct mmu_state *state)
   resolution = state_read_mem_byte(&state->data[16*2+4*9]);
 }
 
-void shifter_build_ppm()
-{
-#if 0
-  int x,y,c;
-  char header[80];
-  unsigned char frame[384*288*3];
-
-  c = 0;
-
-  for(y=0;y<288;y++) {
-    for(x=0;x<384;x++) {
-      frame[c*3+0] = rgbimage[((y+12)*512+x+32)*6+2];
-      frame[c*3+1] = rgbimage[((y+12)*512+x+32)*6+1];
-      frame[c*3+2] = rgbimage[((y+12)*512+x+32)*6+0];
-      c++;
-    }
-  }
-
-  sprintf(header, "P6\n%d %d\n255\n", 384, 288);
-  if(write(ppm_fd, header, strlen(header)) != strlen(header))
-    WARNING(write);
-  if (write(ppm_fd, frame, 384*288*3) != 384*288*3)
-    WARNING(write);
-#endif
-}
-
 void shifter_init()
 {
   struct mmu *shifter;
@@ -327,27 +285,11 @@ void shifter_init()
   hsynccnt = res.hblsize+16; /* Offset for first hbl interrupt */
 
   mmu_register(shifter);
-
-  if(ppmoutput) {
-    ppm_fd = open("ostis.ppm", O_WRONLY|O_CREAT|O_TRUNC, 0644);
-  }
-  framecnt = 0;
-}
-
-int64_t last_vsync_ticks = 0;
-
-static int64_t usec_count() {
-  static struct timeval tv;
-
-  gettimeofday(&tv,NULL);
-  return tv.tv_sec * 1000000 + tv.tv_usec;
 }
 
 void shifter_do_interrupts(struct cpu *cpu, int noint)
 {
   long tmpcpu;
-  int64_t current_ticks = 0;
-  int64_t remaining = 0;
 
   tmpcpu = cpu->cycle - lastcpucnt;
   if(tmpcpu < 0) tmpcpu += MAX_CYCLE;
@@ -363,27 +305,7 @@ void shifter_do_interrupts(struct cpu *cpu, int noint)
     vsynccnt += res.screen_cycles;
     linenum = 0;
     hsynccnt += res.hblsize;
-    if(ppmoutput) {
-      shifter_build_ppm();
-    }
-    if(vsync_delay) {
-      current_ticks = usec_count();
-      remaining = current_ticks - last_vsync_ticks;
-      while(remaining < 20000) {
-	//	printf("DEBUG: Remaining: %ld\n", remaining);
-	current_ticks = usec_count();
-	remaining = current_ticks - last_vsync_ticks;
-      }
-      last_vsync_ticks = current_ticks;
-    }
-    screen_swap(SCREEN_NORMAL);
-    vbl_triggered = 1;
-    framecnt++;
-    if((framecnt&0x3f) == 0) {
-      current_ticks = usec_count();
-      usecs_per_framecnt_interval = current_ticks - last_framecnt_usec;
-      last_framecnt_usec = current_ticks;
-    }
+
   }
 
   /* HBL Interrupt */
@@ -391,7 +313,6 @@ void shifter_do_interrupts(struct cpu *cpu, int noint)
     hblpre = res.hblpre;
     hblscr = res.hblscr;
     hsynccnt += res.hblsize;
-    hbl_triggered = 1;
     cpu_set_interrupt(IPL_HBL, IPL_NO_AUTOVECTOR); /* This _should_ work, but probably won't */
   }
 
@@ -401,23 +322,6 @@ void shifter_do_interrupts(struct cpu *cpu, int noint)
 int shifter_get_vsync()
 {
   return res.screen_cycles-vsynccnt;
-}
-
-int shifter_framecnt(int c)
-{
-  if(c == -1) {
-    framecnt = 0;
-  }
-  return framecnt;
-}
-
-float shifter_fps()
-{
-  if(usecs_per_framecnt_interval) {
-    return 1000000*64.0/usecs_per_framecnt_interval;
-  } else {
-    return 0;
-  }
 }
 
 static int shift_out(WORD *data)
