+++ shifter.h
@@ -22,7 +22,6 @@ void shifter_init();
 void shifter_do_interrupts(struct cpu *, int);
 void shifter_print_status();
 int shifter_get_vsync();
-int shifter_framecnt(int);
 float shifter_fps();
 void shifter_load(WORD);
 void shifter_border(void);
